cleanup: rename files to follow convention
